backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.31.0
yt-dlp>=2025.1.0
pydub==0.25.1
python-multipart==0.0.18
aiofiles==24.1.0
Pillow==11.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
